+++ Classes/Event/MiniGameEvent.cpp
@@ -16,7 +16,6 @@
 #include "Layers/Dungeon/SelectEventLayer.h"
 #include "Layers/Dungeon/PasswordEventLayer.h"
 #include "Layers/Message/CharacterMessagelayer.h"
-#include "Layers/Menu/SaveDataSelector.h"
 
 #include "Datas/Message/CharacterMessageData.h"
 
@@ -338,38 +337,5 @@ void StopCountEvent::run()
     DungeonSceneManager::getInstance()->releaseStopWatch();
     DungeonSceneManager::getInstance()->getScene()->getCountDownDisplay()->slideOut();
 
-    this->setDone();
-}
-
-#pragma mark -
-#pragma mark DisplaySaveMenu
-
-bool DisplaySaveMenu::init(rapidjson::Value& json)
-{
-    if (!GameEvent::init(json)) return false;
-    
-    _saveMenu = SaveDataSelector::create(true);
-    _saveMenu->setVisible(false);
-    _saveMenu->onSaveDataSelectCancelled = CC_CALLBACK_0(DisplaySaveMenu::onExitedSaveMenu, this);
-    
-    return true;
-}
-
-void DisplaySaveMenu::run()
-{
-    DungeonSceneManager::getInstance()->getScene()->addChild(_saveMenu, Priority::SELECT_LAYER);
-    SoundManager::getInstance()->playSE(Resource::SE::TITLE_ENTER);
-    _saveMenu->show();
-}
-
-void DisplaySaveMenu::onExitedSaveMenu()
-{
-    SoundManager::getInstance()->playSE(Resource::SE::BACK);
-    _saveMenu->runAction(
-        Sequence::createWithTwoActions(
-            CallFunc::create([this](){this->_saveMenu->hide();}),
-            DelayTime::create(0.3f)
-        )
-    );
     this->setDone();
 }
