+++ Classes/Event/MiniGameEvent.h
@@ -13,8 +13,6 @@
 
 #include "Datas/Message/CharacterMessageData.h"
 
-class SaveDataSelector;
-
 // 連打イベント
 class ButtonMashingEvent : public GameEvent
 {
@@ -100,22 +98,6 @@ private:
     ~StopCountEvent() {FUNCLOG};
     virtual bool init(rapidjson::Value& json) override;
     virtual void run() override;
-};
-
-
-// DisplaySaveイベント
-class DisplaySaveMenu : public GameEvent
-{
-public:
-    CREATE_FUNC_WITH_PARAM(DisplaySaveMenu, rapidjson::Value&);
-private:
-    SaveDataSelector* _saveMenu { nullptr };
-private:
-    DisplaySaveMenu() { FUNCLOG };
-    ~DisplaySaveMenu() { FUNCLOG };
-    virtual bool init(rapidjson::Value& json) override;
-    virtual void run() override;
-    void onExitedSaveMenu();
 };
 
 #endif /* MiniGameEvent_h */
