+++ Classes/Event/ModalLayerEvent.h
@@ -91,4 +91,20 @@ private:
     virtual void run() override;
 };
 
+// DisplaySaveイベント
+class SaveDataSelector;
+class DisplaySaveMenu : public GameEvent
+{
+public:
+    CREATE_FUNC_WITH_PARAM(DisplaySaveMenu, rapidjson::Value&);
+private:
+    SaveDataSelector* _saveMenu { nullptr };
+private:
+    DisplaySaveMenu() { FUNCLOG };
+    ~DisplaySaveMenu() { FUNCLOG };
+    virtual bool init(rapidjson::Value& json) override;
+    virtual void run() override;
+    void onExitedSaveMenu();
+};
+
 #endif /* defined(__LastSupper__MessageEvent__) */
